requirements.txt
# Tidal DL CLI - Python Dependencies
tidalapi>=0.8.10
requests>=2.31.0
rich>=13.7.0
click>=8.1.7
python-dotenv>=1.0.0
mutagen>=1.47.0
aiohttp>=3.9.0
aiofiles>=23.2.1
